refactor(models): add interface definition in its own file
