+++ package.json
@@ -8,8 +8,8 @@
   "scripts": {
     "build": "tsc",
     "commit": "git-cz",
-    "develop": "env-cmd -f ./config/dev.env ts-node-dev --respawn --no-notify --transpileOnly src/main.ts",
-    "develop:inspect": "env-cmd ./config/dev.env ts-node-dev --inspect-brk --respawn --no-notify --transpileOnly src/main.ts",
+    "dev": "env-cmd -f ./config/dev.env ts-node-dev --respawn --no-notify --transpileOnly src/main.ts",
+    "dev:inspect": "env-cmd ./config/dev.env ts-node-dev --inspect-brk --respawn --no-notify --transpileOnly src/main.ts",
     "lint": "eslint . --ext .js,.ts",
     "prebuild": "rm -rf ./build",
     "semantic-release": "semantic-release",
